% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nepal_pm_alri}
\alias{nepal_pm_alri}
\title{Posterior parameters for PM-ALRI Curve}
\format{A list with two elements
\describe{
  \item{posterior_params}{List containing posterior samples for `beta` and `bS` parameters}
  \item{model_data}{List with minimal model information needed for computing ERC.}
  ...
}}
\source{
(To be added)
}
\usage{
nepal_pm_alri
}
\description{
A list containing the posterior samples and spline information needed
for computing values from the exposure-response curve for PM and ALRIs
in children.
}
\details{
The parameters are output from a model fit via the `bercs` package.
}
\keyword{datasets}
